# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: resource not found
        type: string
    type: object
...
  ? github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_UserResponse
  : properties:
      _links:
        $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginationLinks'
      data:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserResponse'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.UserCreateRequest:
    properties:
      active:
        example: true
        type: boolean
      email:
        example: john@example.com
        maxLength: 255
        type: string
      name:
        example: John Doe
        maxLength: 255
        type: string
      password:
        example: secret123
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  github_com_eenemeene_kitamanager-go_internal_models.UserResponse:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      email:
        example: john@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.Group'
        type: array
      id:
        example: 1
        type: integer
      is_superadmin:
        example: false
        type: boolean
      last_login:
        example: "2024-01-15T10:30:00Z"
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  github_com_eenemeene_kitamanager-go_internal_models.UserUpdateRequest:
    properties:
      active:
        example: false
        type: boolean
      email:
        example: john.updated@example.com
        maxLength: 255
        type: string
      name:
        example: John Doe Updated
        maxLength: 255
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@kitamanager.example.com
    name: API Support
  description: REST API for managing Users, Groups, and Organizations with RBAC support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: KitaManager API
  version: "1.0"
paths:
...
  /api/v1/organizations/{orgId}/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users who are members of any group in the
        specified organization
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      - description: Search by name or email (case-insensitive)
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users in an organization
      tags:
      - users
      - organizations
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all users
      parameters:
      - description: Search by name or email (case-insensitive)
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.PaginatedResponse-github_com_eenemeene_kitamanager-go_internal_models_UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /api/v1/users/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /api/v1/users/{userId}/groups:
    post:
      consumes:
      - application/json
      description: Add a user to a group with a specific role (admin, manager, or
        member)
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Group ID and role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.AddUserToGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add user to group with role
      tags:
      - users
  /api/v1/users/{userId}/groups/{groupId}:
    delete:
      consumes:
      - application/json
      description: Remove a user from a group
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove user from group
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's role within a specific group
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UpdateUserGroupRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user's role in group
      tags:
      - users
  /api/v1/users/{userId}/memberships:
    get:
      consumes:
      - application/json
      description: Get all group memberships for a user with effective organization
        roles
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserMembershipsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's group memberships
      tags:
      - users
  /api/v1/users/{userId}/organizations:
    post:
      consumes:
      - application/json
      description: Add a user to an organization's default group with member role
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Organization ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.AddToOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add user to organization
      tags:
      - users
  /api/v1/users/{userId}/organizations/{orgId}:
    delete:
      consumes:
      - application/json
      description: Remove a user from all groups in an organization
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove user from organization
      tags:
      - users
  /api/v1/users/{userId}/superadmin:
    put:
      consumes:
      - application/json
      description: Set or unset a user's superadmin status. Requires superadmin access.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Superadmin status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.SetSuperAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_eenemeene_kitamanager-go_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set user's superadmin status
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"